+++ requirements.txt
@@ -3,7 +3,9 @@ beautifulsoup4==4.11.1
 selenium==4.1.5
 urllib3==1.26.9
 tqdm==4.64.0
-webdriver-manager==3.5.4
+# Breaking changes were added in 3.6.0
+# See https://github.com/SergeyPirogov/webdriver_manager/pull/377
+webdriver-manager>=3.5.2,<=3.5.4
 
 # packages for contributing
 autopep8==1.6.0
